+++ test/pinput_test.dart
@@ -18,11 +18,21 @@ void main() {
     final length = 4;
     final focusNode = FocusNode();
     final defaultTheme = PinTheme(decoration: BoxDecoration(color: Colors.red));
-    final focusedTheme = defaultTheme.copyDecorationWith(color: Colors.greenAccent.withOpacity(.9));
-    final submittedTheme = defaultTheme.copyDecorationWith(color: Colors.greenAccent.withOpacity(.8));
-    final followingTheme = defaultTheme.copyDecorationWith(color: Colors.greenAccent.withOpacity(.7));
-    final disabledTheme = defaultTheme.copyDecorationWith(color: Colors.greenAccent.withOpacity(.6));
-    final errorTheme = defaultTheme.copyDecorationWith(color: Colors.greenAccent.withOpacity(.5));
+    final focusedTheme = defaultTheme.copyDecorationWith(
+      color: Colors.greenAccent.withOpacity(.9),
+    );
+    final submittedTheme = defaultTheme.copyDecorationWith(
+      color: Colors.greenAccent.withOpacity(.8),
+    );
+    final followingTheme = defaultTheme.copyDecorationWith(
+      color: Colors.greenAccent.withOpacity(.7),
+    );
+    final disabledTheme = defaultTheme.copyDecorationWith(
+      color: Colors.greenAccent.withOpacity(.6),
+    );
+    final errorTheme = defaultTheme.copyDecorationWith(
+      color: Colors.greenAccent.withOpacity(.5),
+    );
 
     await tester.pumpApp(
       Pinput(
@@ -39,7 +49,9 @@ void main() {
 
     void _testState(int count, PinTheme theme) {
       expect(
-        find.byWidgetPredicate((w) => w is AnimatedContainer && w.decoration == theme.decoration),
+        find.byWidgetPredicate(
+          (w) => w is AnimatedContainer && w.decoration == theme.decoration,
+        ),
         findsNWidgets(count),
       );
     }
@@ -82,7 +94,8 @@ void main() {
     _testState(0, disabledTheme);
   });
 
-  testWidgets('Should properly handle focused state', (WidgetTester tester) async {
+  testWidgets('Should properly handle focused state',
+      (WidgetTester tester) async {
     final focusNode = FocusNode();
     final defaultTheme = PinTheme(decoration: BoxDecoration());
     final focusedTheme = defaultTheme.copyDecorationWith(color: Colors.red);
@@ -102,7 +115,10 @@ void main() {
     expect(find.text('|'), findsOneWidget);
 
     expect(
-      find.byWidgetPredicate((w) => w is AnimatedContainer && w.decoration == focusedTheme.decoration),
+      find.byWidgetPredicate(
+        (w) =>
+            w is AnimatedContainer && w.decoration == focusedTheme.decoration,
+      ),
       findsOneWidget,
     );
   });
@@ -120,7 +136,8 @@ void main() {
   });
 
   group('onChanged should work properly', () {
-    testWidgets('onChanged should work with controller', (WidgetTester tester) async {
+    testWidgets('onChanged should work with controller',
+        (WidgetTester tester) async {
       String? fieldValue;
       int called = 0;
 
@@ -151,7 +168,8 @@ void main() {
       expect(called, 2);
     });
 
-    testWidgets('onChanged should work with controller', (WidgetTester tester) async {
+    testWidgets('onChanged should work with controller',
+        (WidgetTester tester) async {
       String? fieldValue;
       int called = 0;
       final TextEditingController controller = TextEditingController();
@@ -188,7 +206,8 @@ void main() {
   });
 
   group('onCompleted should work properly', () {
-    testWidgets('onCompleted works without controller', (WidgetTester tester) async {
+    testWidgets('onCompleted works without controller',
+        (WidgetTester tester) async {
       String? fieldValue;
       int called = 0;
 
@@ -272,7 +291,8 @@ void main() {
     expect(tapCount, 3);
   });
 
-  testWidgets('onTap is not called, field is disabled', (WidgetTester tester) async {
+  testWidgets('onTap is not called, field is disabled',
+      (WidgetTester tester) async {
     int tapCount = 0;
 
     await tester.pumpApp(